tg: detach 'for-review' from TopGit
If you use TopGit internally to manage your different commits you want
to integrate on top of other MPTCP changes (on top of t/upstream or
export), you can now base your tree on top of 'for-review' branch (you
can also use it as target for 'git-review' or keep 't/upstream', that
will be the same from a Gerrit point of view): this new branch
'for-review' is like 't/upstream' (top of the tree) but without TopGit
files: Gerrit will see when changes are integrated in the tree (not like
'export') because Change-Id are preserved and mark changes as merged
automatically when it is in the tree. Also with this new branch, you
don't have any .top* files on this branch.

In other words, use this branch as base and target for new changes.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Change-Id: I2a085ebaab81c0079944fb098ceeed5b2ceb0bab